wip: move form into a fragment
