+++ app/src/main/res/menu/bottom_nav.xml
@@ -3,7 +3,7 @@
 
     <item
         android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_refresh"
+        android:icon="@drawable/ic_home_black_24dp"
         android:title="Home" />
 
     <item
